+++ modules/fabric_1.18.2/src/main/java/com/armilp/ezvcsurvival/EZVCSurvival.java
@@ -4,8 +4,9 @@ package com.armilp.ezvcsurvival;
 import com.armilp.ezvcsurvival.config.VoiceConfig;
 import com.mojang.logging.LogUtils;
 
-import fuzs.forgeconfigapiport.api.config.v2.ForgeConfigRegistry;
+
 import net.fabricmc.api.ModInitializer;
+import net.minecraftforge.api.ModLoadingContext;
 import net.minecraftforge.fml.config.ModConfig;
 import org.slf4j.Logger;
 
@@ -15,6 +16,6 @@ public class EZVCSurvival implements ModInitializer {
 
     @Override
     public void onInitialize() {
-        ForgeConfigRegistry.INSTANCE.register(MOD_ID, ModConfig.Type.COMMON, VoiceConfig.CONFIG);
+        ModLoadingContext.registerConfig(MOD_ID, ModConfig.Type.COMMON, VoiceConfig.CONFIG);
     }
 }