+++ modules/fabric_1.21.4/src/main/java/com/armilp/ezvcsurvival/FollowVoiceGoal.java
@@ -3,6 +3,7 @@ package com.armilp.ezvcsurvival;
 import net.minecraft.entity.ai.goal.Goal;
 import net.minecraft.entity.mob.MobEntity;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.math.BlockPos;
 
 import java.util.EnumSet;
@@ -13,6 +14,7 @@ public class FollowVoiceGoal extends Goal {
     private final double speedModifier;
     private final int voiceDetectionRange;
     private PlayerEntity targetPlayer;
+    private ServerWorld world;
     private final double threshold;
     private BlockPos targetSoundPosition;
     private long timePlayerInRange;
@@ -87,7 +89,7 @@ public class FollowVoiceGoal extends Goal {
             if (currentTime - lastAttackTime >= attackCooldown) {
                 mob.getNavigation().stop();
                 mob.swingHand(mob.preferredHand); // Animación del ataque
-                mob.tryAttack(targetPlayer);   // Realiza el ataque
+                mob.tryAttack(world, targetPlayer);   // Realiza el ataque
                 lastAttackTime = currentTime;
             }
             return;